feat: implement logical clock (#6)
* feat: implement logical clock

Previously, the benchmark runner uses physical time to track the
benchmark progress. This is not ideal because the physical time keeps
moving forward even when the benchmark is paused. Also the throughput
calculation is not accurate when there are pauses. This commit introduces
a logical clock that can be paused and resumed to improve the pause
functionality.

Signed-off-by: Wenxuan Zhang <wenxuangm@gmail.com>